qemu-arm: Reserve memory region for MM_COMMUNICATE buffer
The MM_COMMUNICATE call needs a buffer which can be used to share data
between u-boot and the StandaloneMm(StMm) image running in the Secure
world.

The qemu virt platform does not have a trustzone controller that can
be used to configure such a buffer, but it has a fixed memory map for
secure and non-secure memory regions. Carve out a portion of the
non-secure DRAM to be used as a buffer for MM_COMMUNICATE. Do this by
restricting the size of DRAM to 2GB, and reserving 1GB for the buffer.

This also requires the qemu to be invoked with -m 3G flag, to provide
for a total of 3GB RAM memory.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>