Add support for fetching uefi variable via MM interface
Add support for fetching uefi variable data over Management Mode(MM)
interface. This is just a proof of concept to show how the GetVariable
functionality can be implemented in u-boot to get uefi variables from
the Secure world running StandaloneMm(StMm) image.

The functionality can be invokes via a u-boot command
'get_var_mm'. Changes are also required in tf-a and StMm code for this
command to work.

Signed-off-by: Sughosh Ganu <sughosh.ganu@linaro.org>